+++ addon/components/file-browser/component.js
@@ -40,12 +40,11 @@ export default Ember.Component.extend(Analytics, {
     selectedFile: null,
     node: null,
     nodeTitle: null,
-    newNodeType: null,
+    newProject: false,
     projectSelectState: 'main',
-    willCreateComponent: false,
-    willMoveToNode: false,
     isInputInvalid: true,
     nodeLink: Ember.computed.alias('node.links.html'),
+    isMoving: false,
 
     init() {
         this._super(...arguments);
@@ -384,8 +383,6 @@ export default Ember.Component.extend(Analytics, {
         closeModal() {
             if (this.get('modalOpen') == 'move') {
                 this.set('projectSelectState', 'main');
-                this.set('willCreateComponent', false);
-                this.set('willMoveToNode', false);
             }
             this.set('modalOpen', false);
         },
@@ -406,59 +403,32 @@ export default Ember.Component.extend(Analytics, {
         },
         checkNodeTitleKeypress(value) {
             this.set('nodeTitle', value);
-            if (this.get('nodeTitle')) {
-                this.set('isInputInvalid', false);
-            } else {
-                this.set('isInputInvalid', true);
-            }
+            this.set('isInputInvalid', Ember.isBlank(value));
         },
         changeProjectSelectState(state) {
             this.set('projectSelectState', state);
-            this.set('willCreateComponent', false);
-            this.set('willMoveToNode', false);
-            if (!this.get('isInputInvalid')) {
-                this.set('isInputInvalid', true);
-            }
-        },
-        updateCreateOrMoveNode(state) {
-            if (state == 'create') {
-                this.set('willCreateComponent', true);
-                this.set('willMoveToNode', false);
-                this.set('isInputInvalid', true);
-            } else {
-                this.set('willCreateComponent', false);
-                this.set('willMoveToNode', true);
-                this.set('isInputInvalid', false);
-            }
+            this.set('isInputInvalid', true);
         },
         setSelectedNode(node, isChild) {
             this.set('node', node);
             this.set('isChildNode', isChild);
+            this.set('isInputInvalid', false);
         },
         setMoveFile() {
+            this.set('isMoving', true);
             let selectedItem = this.get('selectedItems.firstObject');
             let title = this.get('nodeTitle');
 
             if (this.get('projectSelectState') == 'newProject') {
-                this.set('newNodeType', 'project');
+                this.set('newProject', true);
                 this._createProject(title).then(project => {
                     this.set('node', project);
                     this._addNewNodeToList(this.get('user'), project);
                     this._moveFile(selectedItem, project);
                 });
             } else if (this.get('projectSelectState') == 'existingProject') {
-                if (this.get('willCreateComponent')) {
-                    this.set('newNodeType', 'component');
-                    this._createComponent(title).then(component => {
-                        this.set('node', component);
-                        this._addNewNodeToList(this.get('user'), component);
-                        this._moveFile(selectedItem, component);
-                    });
-                } else if (this.get('willMoveToNode')) {
-                    this._moveFile(selectedItem, this.get('node'));
-
-                    this.set('newNodeType', null);
-                }
+                this._moveFile(selectedItem, this.get('node'));
+                this.set('newProject', false);
             }
         },
     },
@@ -472,13 +442,6 @@ export default Ember.Component.extend(Analytics, {
                 this.get('i18n').t('eosf.components.moveToProject.couldNotCreateProject')
             ));
     },
-    _createComponent(title) {
-        return this.get('node')
-            .addChild(title, null, '', true)
-            .catch(() => this.get('toast').error(
-                this.get('i18n').t('eosf.components.moveToProject.couldNotCreateComponent')
-            ));
-    },
     _moveFile(item, node) {
         item.move(node)
             .then(() => {
@@ -490,11 +453,11 @@ export default Ember.Component.extend(Analytics, {
                 this.set('modalOpen', 'successMove');
                 this.set('projectSelectState', 'main');
                 this.set('willCreateComponent', false);
-                this.set('willMoveToNode', false);
                 this.send('track', 'move', 'track', 'Quick Files - Move to project');
             })
             .catch(() => this.get('toast').error(
                 this.get('i18n').t('eosf.components.moveToProject.couldNotMoveFile')
-            ));
+            ))
+            .then(() => this.set('isMoving', false));
     },
 });