+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added "choose your custom citation" section to citation-widget
 - `facebookAppId` field to `preprint-provider` model
+- Add ElasticSearch preference (ES) key to preprints ES search requests for reproducible results ordering.
 
 ## [0.14.0] - 2018-02-07
 ### Added
@@ -24,7 +25,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Move function for `file` model
 - `ember-collapsable-panel` and `ember-power-select` packages
 - Styling on `project-selector` buttons that are selected
-- Add ElasticSearch preference (ES) key to preprints ES search requests for reproducible results ordering.
 
 ### Changed
 - Node `addChild` model function to create public child elements