+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added "choose your custom citation" section to citation-widget
 - `facebookAppId` field to `preprint-provider` model
+- Add ElasticSearch preference (ES) key to preprints ES search requests for reproducible results ordering.
 
 ## [0.14.0] - 2018-02-07
 ### Added