+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Ability to create new components when moving files.
+
+## [Unreleased 0.15.0]
+### Added
+- Added "choose your custom citation" section to citation-widget
+- `facebookAppId` field to `preprint-provider` model
+- Add ElasticSearch preference (ES) key to preprints ES search requests for reproducible results ordering.
 
 ## [0.14.0] - 2018-02-07
 ### Added